requirements.txt
torch==1.13.1
tensorboardX==2.6
matplotlib==3.7.0
PyYAML==6.0
tqdm==4.64.1
h5py==3.8.0
protobuf==3.20.0
hydra-core==1.3.2
midi2audio==0.1.1
pypianoroll==1.0.4
scipy==1.10.1
tensorboard==2.12.0
